refactor: adjusting class responisblities
