+++ manifest.php
@@ -23,6 +23,7 @@ use oat\taoMediaManager\scripts\install\SetMediaManager;
 use oat\taoMediaManager\model\user\TaoAssetRoles;
 use oat\tao\model\accessControl\func\AccessRule;
 use oat\taoMediaManager\scripts\install\SetRolesPermissions;
+use oat\taoMediaManager\scripts\install\SetupMiddlewares;
 
 $extpath = __DIR__ . DIRECTORY_SEPARATOR;
 $taopath = __DIR__ . DIRECTORY_SEPARATOR . 'tao' . DIRECTORY_SEPARATOR;
@@ -161,6 +162,7 @@ return [
         'php' => [
             SetMediaManager::class,
             SetRolesPermissions::class,
+            SetupMiddlewares::class,
         ]
     ],
     'update' => 'oat\\taoMediaManager\\scripts\\update\\Updater',
